fix type of variables. Int to float

+++ moncenterlib/gnss/gnss_time_series.py
@@ -47,21 +47,21 @@ def parse_pos_file(path2file: str, sep: str | None = None) -> tuple[dict[str, li
                     row.pop(3)  # date dms dms m
                     row.pop(3)  # date dms dms
                 else:
-                    row[1] = float(row[1]) # coord 1
-                    row[2] = float(row[2]) # coord 2
-                
-                row[3] = float(row[3]) # coord 3
-                row[4] = int(row[4]) # Q
-                row[5] = int(row[5]) # ns
-                
-                row[6] = int(row[6]) # sd coord 1
-                row[7] = int(row[7]) # sd coord 2
-                row[8] = int(row[8]) # sd coord 3
-                row[9] = int(row[9]) # sd coord 1 2
-                row[10] = int(row[10]) # sd coord 2 3
-                row[11] = int(row[11]) # sd coord 1 3
-                row[12] = int(row[12]) # age
-                row[13] = int(row[13]) # ratio 
+                    row[1] = float(row[1])  # coord 1
+                    row[2] = float(row[2])  # coord 2
+
+                row[3] = float(row[3])  # coord 3
+                row[4] = int(row[4])  # Q
+                row[5] = int(row[5])  # ns
+
+                row[6] = float(row[6])  # sd coord 1
+                row[7] = float(row[7])  # sd coord 2
+                row[8] = float(row[8])  # sd coord 3
+                row[9] = float(row[9])  # sd coord 1 2
+                row[10] = float(row[10])  # sd coord 2 3
+                row[11] = float(row[11])  # sd coord 1 3
+                row[12] = float(row[12])  # age
+                row[13] = float(row[13])  # ratio
 
                 data.append(row)
